Make changes caused the data to change format. This broke paper. Now fixed.

+++ src/main/java/org/made/neohabitat/Document.java
@@ -138,7 +138,7 @@ public abstract class Document extends HabitatMod {
 					for (int i = 0; i < last_page; i++) {
 						Iterator<Object> chars = ((JSONArray) bytePage.next()).iterator();
 						for (int j = 0; chars.hasNext() ; j++) {
-							int c = ((Double) chars.next()).intValue();
+							int c = ((Long) chars.next()).intValue();
 							if (c == 0) {
 								break;
 							}
